Move spec for team participation in tournament
