+++ spec/controllers/tournaments_controller_spec.rb
@@ -118,55 +118,4 @@ RSpec.describe TournamentsController, type: :controller do
       is_expected.to redirect_to tournaments_path
     end
   end
-
-  describe "#add_team" do
-    subject(:adding) { post :add_team, params: params }
-    let(:tournament) { create(:tournament, owner: current_user) }
-    let(:members) { create_list(:user, 3) }
-    let(:params) { { tournament_id: tournament.id, team: team_params } }
-    let(:team_params) { { name: "foo", member_ids: members.pluck(:id) } }
-
-    context "when the tournament is open" do
-      context "when adding a unique team" do
-        it "adds the team" do
-          expect { adding }.to change { tournament.teams.count }.by(1)
-        end
-      end
-
-      context "when adding a duplicate team" do
-        before { create(:team, name: "bar", members: members) }
-
-        it "adds a reused the team" do
-          expect { adding }
-            .to change { tournament.reload.teams.pluck(:name) }
-            .to include("bar")
-        end
-      end
-    end
-
-    context "when the tournament is not open" do
-      before { tournament.started! }
-
-      it { expect { adding }.to raise_error(Pundit::NotAuthorizedError) }
-    end
-  end
-
-  describe "#remove_team" do
-    subject(:removing) { post :remove_team, params: params }
-    let(:params) { { tournament_id: tournament.id, team_id: team.id } }
-    let(:tournament) { create(:tournament, teams: [team], owner: current_user) }
-    let(:team) { create(:team) }
-
-    context "when the tournament is open" do
-      it "removes the team" do
-        expect { removing }.to change { tournament.teams.count }.by(-1)
-      end
-    end
-
-    context "when the tournament is not open" do
-      before { tournament.started! }
-
-      it { expect { removing }.to raise_error(Pundit::NotAuthorizedError) }
-    end
-  end
 end
